Checkpoint 7.1 moved stuff. Drafted scoped version of allocator, but it's hard to apply(yet?), so unused
