+++ rts/Rendering/Models/MatricesMemStorage.cpp
@@ -2,5 +2,5 @@
 
 MatricesMemStorage::MatricesMemStorage()
 {
-	spa = std::make_unique<spring::StablePosAllocator<CMatrix44f>>(INIT_NUM_ELEMS);
+	spa = std::make_unique<StablePosAllocator<CMatrix44f>>(INIT_NUM_ELEMS);
 }
